adds ability to generate list of sprint issues
I'm not quite sure what the actual business term should be here, but
I'm trying to distinguish between the rows in the Jira export, the
Stories they represent (which can span multiple sprints), and then the
items that represent said story in each sprint. This is all still a
bit jumbled for me, but I think it's at least directionally right.