moves sprint_issues into a method
I'm going to work on a second interface for this report, and that will
basically be a view onto this same data structure, so I need to make
that structure more generally available.

Note that I'm not currently certain if this should be public or
private. My gut says private, since no one ought to ask the Report
about it's sprint_issues but should instead only ask for a
representation of those. But also I am about to make a change that
requires me to test sprint_issues, so I'm torn. I suppose there is
another object lurking about here, but I'll wait til I know more
before I try to expose it.