makes as_sprint_issues return hash, not array
I'm honestly not entirely sure how this needs to evolve in the future,
but when I was starting to think about a different type of report I
realized that insisting on a certain order of array elements. I think
eventually it might make sense to get away from using a primitive
object here and actually have a SprintIssue object, but until I have a
better sense of what that would look like, I'll stick with a hash.